Implementation of Clustering Algorithm
These are the different/independent variables that have been given. Here we are talking about Unsupervised learning so there is no Y to predict. Based on this data I have to identify the pattern that what does it tell which cereals are similar to each other. May be am a dietician and I have to recommend someone based on your fitness what cereals are good for you. Some cereals give good protein, carbs, etc.

Here my goal is to bunch a group of rows based on their similarity in a cluster, where observations in one cluster are similar to each other but dissimilar for the different clusters. Here, I am aggregating "name" to do clustering don't know, "Shelf" not needed because first, I should know on what cereals together am doing a clustering, "rating" not needed because it's not a regression problem...in this case, we can say based on these values what would be the rating for that cereals. So, just I don't want to lose information am combining all three attributes as labels so that cereals which have this rating, which is on this shelf, with this name are part of one cluster.